Move assessment creation into staff base
